back-end/requirements.txt
asgiref==3.8.1
coverage==7.5.1
ddt==1.7.2
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-rest-passwordreset==1.4.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
iniconfig==2.0.0
Markdown==3.6
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.0
Pillow==10.3.0
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.1
pytest-cov==5.0.0
pytz==2024.1
sqlparse==0.5.0
tomli==2.0.1
stripe==9.6.0
